problems resolved
