+++ src/index.js
@@ -3,12 +3,12 @@ import { chatGrupal } from './views/groupChat.js';
 // ... import other views
 import { setRootEl, setRoutes, onURLChange } from './router.js';
 //import dataset from './data/dataset.js';
-import { filterData, sortData, computeStats } from './lib/dataFunctions.js'
-import { renderItems } from './components/cards.js'
-import data from './data/dataset.js'
-import { main } from './components/main.js'
-import { Header } from './components/header.js';
-import { Footer } from './components/footer.js';
+//import { filterData, sortData, computeStats } from './lib/dataFunctions.js'
+//import { renderItems } from './components/cards.js'
+//import data from './data/dataset.js'
+//import { main } from './components/main.js'
+//import { Header } from './components/header.js';
+//import { Footer } from './components/footer.js';
 
 
 
@@ -37,70 +37,19 @@ window.addEventListener('DOMContentLoaded', () => {
   const rootElement = document.querySelector('#root');
   setRootEl(rootElement);
   onURLChange(); 
-  rootElement.appendChild(Header());
-  rootElement.appendChild(main());
+  //rootElement.appendChild(Header());
+  //rootElement.appendChild(main());
 
-  const platformSelect = document.querySelector('#platform');
+  /*const platformSelect = document.querySelector('#platform');
   const sortBySelect = document.querySelector('#sortBy');
   const buttonReset = document.querySelector('#buttonClear');
   const buttonStats = document.querySelector('#buttonStats');
   const resultsContainer = document.querySelector('#results');
 
-  const originalData = [...data];
+  const originalData = [...data];*/
 
   //console.log({ platformSelect, sortBySelect, buttonReset, buttonStats, resultsContainer }); 
 
-  if (platformSelect && sortBySelect && buttonReset && buttonStats && resultsContainer) {
-
-    const renderFilteredData = () => {
-      const platform = platformSelect.value;
-      const sortByOption = sortBySelect.value.split('-');
-      const sortBy = sortByOption[0];
-      const sortOrder = sortByOption[1];
-
-      let filteredData = data;
-      if (platform) {
-        filteredData = filterData(filteredData, 'streamingPlatform', platform);
-      }
-
-      if (sortBy && sortOrder) {
-        filteredData = sortData(filteredData, sortBy, sortOrder);
-      }
-
-      //console.log('Filtered Data:', filteredData); 
-      resultsContainer.innerHTML = '';
-      resultsContainer.appendChild(renderItems(filteredData));
-    };
-
-    platformSelect.addEventListener('change', () => {
-      //console.log('Platform changed:', platformSelect.value);
-      renderFilteredData();
-    });
-
-    sortBySelect.addEventListener('change', () => {
-      // console.log('Sort by changed:', sortBySelect.value);
-      renderFilteredData();
-    });
-
-    buttonReset.addEventListener('click', () => {
-      //console.log('Reset button clicked');
-      platformSelect.selectedIndex = 0;
-      sortBySelect.selectedIndex = 0;
-      resultsContainer.innerHTML = '';
-      resultsContainer.appendChild(renderItems(originalData));
-    });
-
-    buttonStats.addEventListener('click', () => {
-      //console.log('Stats button clicked');
-      const stats = computeStats(data);
-      //console.log('Stats:', stats); // Debugging
-      resultsContainer.innerHTML = `
-        <h4><span class="highlight">${stats.avgYears.toFixed(2)} años</span> promedia una transmisión.</h4>
-      `;
-    });
-
-    resultsContainer.appendChild(renderItems(data));
-  }
-
-  rootElement.appendChild(Footer());
+  
+  //rootElement.appendChild(Footer());
 });
